Improved the drag dro pmanager Improvd the command events Load the saved file

+++ warcommands-app/src/warcommands/commands-panel/domain/command-component/services/command-component-manager.service.ts
@@ -27,8 +27,11 @@ export class CommandComponentManagerService {
 
     removeComponent(commandId: string) {
         const component: ComponentRef<any> = this.commandComponentRepositoryService.findByCommandId(commandId);
-        component.destroy();
-        this.commandComponentRepositoryService.remove(commandId);
+
+        if (component) {
+            component.destroy();
+            this.commandComponentRepositoryService.remove(commandId);
+        }
     }
 
 }
